requirements.txt
Pillow>=8.0.0
python-vlc>=3.0.0